requirements.txt
streamlit
instaloader
pandas
openpyxl
pillow